% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWD_MOSMIX.R
\name{get_meta_info}
\alias{get_meta_info}
\alias{print.dwdmeta}
\title{Extract DWD MOSMIX Forecast Meta Information}
\usage{
get_meta_info(doc)

\method{print}{dwdmeta}(x, ...)
}
\arguments{
\item{doc}{an \code{XMLInternalDocument} object as returned by
\code{xmlParse} (XML package).}

\item{x}{an object of class \code{dwdmeta} as returned by
\code{\link{get_meta_info}}.}

\item{...}{forwarded to S3 methods. Unused for now.}
}
\value{
Returns an object of class \code{dwdmeta}.
}
\description{
The DWD MOSMOX forecasts are provided as XML files. The "header"
of the XML file contains some information about the MOSMIX model,
initialization time, and so far and so on. This method extracts
the meta information.
}
\details{
Extracts the following information from the XML file:
\itemize{
     \item \code{Issuer} who created these forecasts.
     \item \code{ProductID}
     \item \code{GeneratingProcess} 
     \item \code{IssueTime} the initialization time (most important
           information). Will also be used when creating the output files
           (see \code{\link{write_ascii}}).
     \item \code{ReferencedModel} the NWP models the MOS is based on.
}
}
\examples{
# Latest L-type DWD MOSMIX forecast file for Innsbruck Airport
url <- "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/11120/kml/MOSMIX_L_LATEST_11120.kmz"

# Download and extract file
kmz   <- tempfile("mosmix_demo_", fileext = ".kmz")
check <- download.file(url, kmz)
if ( inherits(check, "try-error") ) stop("Problems downloading the file!")
kml   <- unzip(kmz)

# Parsing the unzipped kml file (XML format)
doc <- XML::xmlParse(kml)

# Extracting meta information
meta <- get_meta_info(doc)
print(meta)

# Remove kmz and kml file
file.remove(kmz, kml)

}
\seealso{
get_station_information, get_meta_info, get_datetime.
}
\author{
Reto Stauffer

Reto Stauffer
}
