% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWD_MOSMIX.R
\name{parameter_description}
\alias{parameter_description}
\title{Load DWD MOS Parameter Description}
\usage{
parameter_description(...)
}
\arguments{
\item{...}{unused.}
}
\value{
Returns a \code{data.frame} with three columns containing
the name of the parameter (\code{name}), the unit (\code{unit}),
and a description of the parameter (\code{description}).
}
\description{
An XML file on the opendata server provides information about
the forecasted elements.
}
\examples{
# Loading information
desc <- parameter_description()
print(head(desc))

}
\author{
Reto Stauffer
}
