% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dwdforecast.R
\name{plot.dwdforecast}
\alias{plot.dwdforecast}
\title{Default Plot for DWD MOSMIX Forecasts}
\usage{
\method{plot}{dwdforecast}(x, ...)
}
\arguments{
\item{x}{object of class \code{dwdforecasts} as returned by
\code{\link{get_forecasts}}.}
}
\description{
A default plot routine to check the data.
}
